requirements.txt
opencv-python==4.1.2
pillow==7.0.0

